Refactoring and Bug Fixes
